# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JamJam Trek API",
        "description": "API documentation for JamJam Trek application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://161.97.167.73:8001",
            "description": "Server"
        }
    ],
    "paths": {
        "/api/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Get list of activities",
                "operationId": "07c619d4683474b5c895505122616066",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter activities by category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Create a new activity",
                "operationId": "572f2d7594f010ca449e8ff341f22c8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "location",
                                    "price",
                                    "currency",
                                    "duration",
                                    "difficulty",
                                    "category",
                                    "is_featured",
                                    "is_active"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Paragliding in Pokhara"
                                    },
                                    "location": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Pokhara"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 85
                                    },
                                    "currency": {
                                        "type": "string",
                                        "maxLength": 10,
                                        "example": "USD"
                                    },
                                    "duration": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "30 minutes"
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Easy"
                                    },
                                    "category": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Aerial"
                                    },
                                    "min_age": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 12
                                    },
                                    "max_participants": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "inclusions": {
                                        "type": "string",
                                        "example": "Equipment, Insurance, Guide"
                                    },
                                    "requirements": {
                                        "type": "string",
                                        "example": "Good health, No fear of heights"
                                    },
                                    "featured_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "gallery_images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "is_featured": {
                                        "type": "boolean"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "season": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "All Year"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Activity created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Activity created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/activities/{id}": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Get a single activity",
                "operationId": "26bc2667123f75faff01558c4e1cc599",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Activities"
                ],
                "summary": "Update an activity",
                "operationId": "b655861c456cc1c1c1bfb63a2dc2fcac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "location": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0
                                    },
                                    "currency": {
                                        "type": "string",
                                        "maxLength": 10
                                    },
                                    "duration": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "category": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "min_age": {
                                        "type": "integer",
                                        "minimum": 0
                                    },
                                    "max_participants": {
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "inclusions": {
                                        "type": "string"
                                    },
                                    "requirements": {
                                        "type": "string"
                                    },
                                    "featured_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "gallery_images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "is_featured": {
                                        "type": "boolean"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "season": {
                                        "type": "string",
                                        "maxLength": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Activity updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Activities"
                ],
                "summary": "Delete an activity",
                "operationId": "70a1e6d0404ea2ee8cc177b860fee18a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}